Removed the extra level of indirection from our "Function" type
