+++ crates/rune-cli/Cargo.toml
@@ -31,7 +31,7 @@ rune-runtime = { path = "../runtime", version = "0.4.0-dev" }
 rune-syntax = { path = "../syntax", version = "0.4.0-dev" }
 rune-wasmer-runtime = { path = "../wasmer-runtime", version = "0.4.0-dev" }
 rune-core = { path = "../rune-core", version = "0.4.0-dev" }
-runicos-base-runtime = { path = "../../images/runicos-base/runtime", version = "0.4.0-dev" }
+runicos-base-runtime = { path = "../../images/runicos-base/runtime", version = "0.4.0-dev", features = ["tflite"] }
 serde = { version = "1.0.125", features = ["derive"] }
 serde_json = "1.0.64"
 structopt = "0.3.21"
