+++ images/runicos-base/runtime/src/lib.rs
@@ -1,5 +1,3 @@
 mod image;
-mod image_v2;
 
 pub use image::{BaseImage, Model};
-pub use image_v2::{BaseImage as BaseImage2, Model as Model2};
